#35 refactor: 이메일 인증 확인 로직 위치 변경 및 MemberService 로직 변경
